Adds the work command
It adds the work command which basically finds a ticket with the
provided name [creates] and checks out a branch with that name.

Now dits will also discover correctly the repository issues is in,
even if it's inside a git submodule.

Added a couple of utility methods, some documentation and
some tests (more to be added!).